Update toolchain: xunit 4/MTP, SharpCompress 0.50.4, .NET 10 dotnet test mode
- LzmaCodec: LzmaStream.Create factory API (0.50.4 removed the ctor);
  output verified byte-compatible raw headerless LZMA (chdman LZMA suites pass)
- dotnet test: global.json test.runner = Microsoft.Testing.Platform (new .NET 10
  experience; VSTest mode is rejected for MTP v2 apps); --logger replaced with -v
- xunit.v3 4.0.0 / runner 4.0.0 / Test.Sdk 18.9.0 / Meziantou.Analyzer 3.0.159
- Repo-wide style modernization (PascalCase codec/map constants, brace style,
  redundant-cast cleanup); docs updated

+++ CHDSharpEncoderTest/CueParserTests.cs
@@ -392,7 +392,7 @@ public class CueParserTests : IDisposable
         void WriteFourCc(string tag) => w.Write(System.Text.Encoding.ASCII.GetBytes(tag));
 
         WriteFourCc("RIFF");
-        w.Write((uint)(36 + dataLength)); // RIFF chunk size
+        w.Write(36 + dataLength); // RIFF chunk size
         WriteFourCc("WAVE");
         WriteFourCc("fmt ");
         w.Write(16u);       // fmt chunk size
